backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
langchain>=0.1.0
langchain-openai>=0.0.5
chromadb>=0.4.18
pypdf2>=3.0.1
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
openai>=1.3.7
numpy>=1.26.0
sentence-transformers>=2.2.2
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
matplotlib>=3.8.0
pandas>=2.1.0
requests>=2.31.0
seaborn>=0.13.0
